code cleanup; add token schema
