+++ fastapi/api/v1/routes/login.py
@@ -1,31 +1,14 @@
 from typing import Optional
 
-from pydantic import BaseModel
+from fastapi import APIRouter, Depends, HTTPException
+from fastapi.security import OAuth2PasswordRequestForm
 
-from fastapi import APIRouter, Depends, HTTPException, Request
-from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
-
-from ...core.config import Settings
 from ...core.util import TokenFactory, build_json_hash
 from ...model import User as UserInDB
+from ..depends import get_current_active_user, get_token_factory
+from ..schema import Token, User
 
 router = APIRouter()
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl='./api/v1/login')
-
-
-class User(BaseModel):
-    username: str
-    email: Optional[str] = None
-    full_name: Optional[str] = None
-    disabled: Optional[bool] = None
-
-    class Config:
-        orm_mode = True
-
-
-async def get_token_factory(request: Request) -> TokenFactory:
-    settings: Settings = request.app.extra['settings']
-    return TokenFactory(secret_key=settings.secret_key)
 
 
 async def verify_password(plain_password: str, hashed_password: str, salt: str) -> bool:
@@ -38,27 +21,7 @@ async def authenticate_user(username: str, password: str) -> Optional[UserInDB]:
         return user
 
 
-async def get_current_user(token: str = Depends(oauth2_scheme), token_factory: TokenFactory = Depends(get_token_factory)) -> UserInDB:
-    to_raise = HTTPException(status_code=401, detail='could not validate token')
-    token_data = await token_factory.parse(token)
-    if token_data:
-        identity = token_data.sub
-        _type, name = identity.split(':')
-        if _type != 'username':
-            raise to_raise
-        user = UserInDB.get(username=name)
-        if not user:
-            raise to_raise
-        return user
-
-
-async def get_current_active_user(current_user: User = Depends(get_current_user)):
-    if current_user.disabled:
-        raise HTTPException(status_code=400, detail='inactive user')
-    return current_user
-
-
-@router.post('/login')
+@router.post('/login', response_model=Token)
 async def login(form_data: OAuth2PasswordRequestForm = Depends(), token_factory: TokenFactory = Depends(get_token_factory)):
     user = await authenticate_user(username=form_data.username, password=form_data.password)
     if not user:
