+++ fastapi/api/v1/schema/__init__.py
@@ -1,1 +1,2 @@
+from .token import Token
 from .user import User
